#92 refactor: login/signup/snslogin/common components/etc
