+++ src/components/common/category/Category.tsx
@@ -1,18 +1,48 @@
 import { useEffect, useState } from "react";
-
-import { CategoryBlock } from "./Category.style";
-import { styled } from "styled-components";
+import {
+  CategoryBlock,
+  ImageWrapper,
+  Overview,
+  TitleContainer,
+  R9dCrew,
+  CategoryContents,
+  CategoryBtns,
+} from "./Category.style";
 import { useNavigate } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 import Skeleton from "../../Skeleton";
 import { setDisplayRemainingComponents } from "../../../redux/modules/Modules";
 import useSearch from "../../../hook/useSearch";
 import LikeButton from "../LikeButton";
+
+type CrewList = {
+  content: string;
+  crewName: string;
+  crewRecruitmentId: number;
+  currentNumber: number;
+  exerciseDate: string;
+  exerciseKind: string;
+  image?: string[];
+  location: string;
+  postDate: number[];
+  title: string;
+};
+
+interface SearchResult {
+  data: CrewList[];
+}
+
+interface RootState {
+  display: {
+    displayRemainingComponents: boolean;
+  };
+}
+
 function Category() {
   const access = localStorage.getItem("Access");
   const refresh = localStorage.getItem("Refresh");
   const navigate = useNavigate();
-  const categories = [
+  const categories: Array<string> = [
     "전체",
     "러닝",
     "웨이트",
@@ -25,20 +55,34 @@ function Category() {
   ];
   const dispatch = useDispatch();
 
-  const [selectedCategory, setSelectedCategory] = useState(""); // 선택된 카테고리
+  const [selectedCategory, setSelectedCategory] = useState<string>("");
 
-  const { searchResult, loading, error, search } = useSearch();
+  const {
+    searchResult,
+    loading,
+    error,
+    search,
+  }: {
+    searchResult: SearchResult["data"];
+    loading: boolean;
+    error: any;
+    search: (keyword: string) => Promise<void>;
+  } = useSearch();
+
+  if (error) {
+    alert("입력하신 키워드를 찾지 못하였습니다.");
+  }
 
   useEffect(() => {
     dispatch(setDisplayRemainingComponents(searchResult.length === 0));
   }, [searchResult, dispatch]);
 
-  const onClickCategory = (pickedCategory) => {
+  const onClickCategory = (pickedCategory: string) => {
     setSelectedCategory(pickedCategory);
     search(pickedCategory);
   };
 
-  const onClickCrew = (itemId) => {
+  const onClickCrew = (itemId: number) => {
     if (access && refresh !== "") {
       navigate(`/crew/${itemId}`);
     } else {
@@ -46,7 +90,7 @@ function Category() {
     }
   };
   const displayRemainingComponents = useSelector(
-    (state) => state.display.displayRemainingComponents
+    (state: RootState) => state.display.displayRemainingComponents
   );
 
   // console.log(selectedCategory);
@@ -57,7 +101,7 @@ function Category() {
     <CategoryBlock>
       {categories.map((category) => {
         return (
-          <button
+          <CategoryBtns
             key={category}
             onClick={() => onClickCategory(category)}
             style={{
@@ -65,14 +109,14 @@ function Category() {
             }}
           >
             {category}
-          </button>
+          </CategoryBtns>
         );
       })}
 
       {!displayRemainingComponents && (
         <CategoryContents>
-          {searchResult?.data?.crewList?.length > 0 &&
-            searchResult.data.crewList.map((post) => {
+          {searchResult?.length > 0 &&
+            searchResult.map((post) => {
               return (
                 <R9dCrew
                   key={post.crewRecruitmentId}
@@ -80,11 +124,7 @@ function Category() {
                 >
                   <ImageWrapper>
                     <img
-                      src={
-                        post.length !== 0 && post.image?.length !== undefined
-                          ? post.image[0]
-                          : ""
-                      }
+                      src={post.image && post.image[0] ? post.image[0] : ""}
                       alt=""
                     />
                   </ImageWrapper>
@@ -103,7 +143,7 @@ function Category() {
                       <LikeButton />
                     </div>
 
-                    <p>{post.exercisekind} / 서울 중구</p>
+                    <p>{post.exerciseKind} / 서울 중구</p>
                   </Overview>
                 </R9dCrew>
               );
@@ -111,78 +151,8 @@ function Category() {
         </CategoryContents>
       )}
       {loading ? <Skeleton /> : ""}
-
-      {error ? alert("입력하신 키워드를 찾지 못하였습니다.") : ""}
     </CategoryBlock>
   );
 }
 
 export default Category;
-const ImageWrapper = styled.div`
-  width: 100%;
-  height: 259px;
-  border-radius: 25px 25px 0 0;
-  background-color: #eeeeee;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  overflow: hidden;
-
-  & img {
-    width: 100%;
-  }
-`;
-
-const Overview = styled.div`
-  width: 100%;
-  height: 70px;
-  padding: 14px 16px;
-  border-radius: 0 0 25px 25px;
-  background-color: #f3f3f3;
-
-  & > div {
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 6px;
-  }
-
-  & > p {
-    font-size: 13px;
-  }
-
-  & > div > span > svg {
-    font-size: 20px;
-  }
-`;
-
-const TitleContainer = styled.div`
-  & > p {
-    display: inline-block;
-    font-size: 18px;
-  }
-
-  & > span {
-    display: inline-block;
-  }
-`;
-
-const R9dCrew = styled.div`
-  width: 398px;
-  height: 232px;
-  cursor: pointer;
-
-  & ${ImageWrapper} {
-    height: 160px;
-  }
-
-  & ${Overview} {
-    height: 72px;
-  }
-`;
-const CategoryContents = styled.section`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 20px;
-  margin-top: 20px;
-`;
